+++ CHANGELOG.md
@@ -51,6 +51,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - add bm25 stats support in index for stastic ranking and dev POC, actually fix tired merge, improve position (add field tags) and core indexing (performance, cleanup) (#112)
    - anubhav will finish this!
 - more config and metrics for rate limits and caching (#113)
+- more metrics and attempts at improvements (#114)
 
 ### Fixed
 - Fix copying body data from buffer (#24)